pyproject.toml
[project]
name = "personal-ai-employee"
version = "0.1.0"
description = "Bronze Tier - Personal AI Employee Foundation with Obsidian vault and Watcher scripts"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Personal AI Employee Team" }
]
dependencies = [
    "specifyplus>=0.0.20.post1",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
